+++ src/components/Bookmark/Skeleton/Card/styled.ts
